refactor: mypage, updatePage 코드 폴더별로 정리
