+++ src/hooks/useUser.ts
@@ -1,23 +1,24 @@
 import { fetchUserInfo } from "@/services/userInfo";
 import { UserInfo } from "@/types/userInfo";
+import { handleApi } from "@/utils/handleApi";
 import { useEffect, useState } from "react";
+import { toast } from "sonner";
 
 export const useUser = () => {
   const [userInfo, setUserInfo] = useState<UserInfo | null>(null);
-  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     const loadUserInfo = async () => {
-      try {
-        const data = await fetchUserInfo();
+      const { data, error } = await handleApi(fetchUserInfo);
+      if (error) {
+        toast.error(error + "회원정보를 불러오지 못했습니다.");
+        return;
+      }
+      if (data) {
         setUserInfo(data);
-      } catch (error) {
-        console.error("Failed to fetch user info", error);
-      } finally {
-        setLoading(false);
       }
     };
     loadUserInfo();
   }, []);
-  return { userInfo, loading };
+  return { userInfo };
 };
