+++ src/services/userInfo.ts
@@ -1,4 +1,4 @@
-import { UserInfo } from "@/types/userInfo";
+import { UserInfo, UserForm } from "@/types/userInfo";
 import api from "@/utils/http-commons";
 
 export const fetchUserInfo = async (): Promise<UserInfo> => {
@@ -8,5 +8,10 @@ export const fetchUserInfo = async (): Promise<UserInfo> => {
 
 export const checkPassword = async (password: string): Promise<boolean> => {
   const { data } = await api.post("/user/passwordcheck", { password });
+  return data;
+};
+
+export const updateUser = async (payload: Partial<UserForm>) => {
+  const { data } = await api.patch("/user", payload);
   return data;
 };
